add fuzzy matching for versions
Extend the existing NodeVersion class with a FuzzyNodeVersion class that
overrides the .match() method behaviour to do a fuzzy match against the
given version.

A FuzzyNodeVersion is created when a given "version" input string is
just a simple string that doesn't map to an existing special case (lts,
current, default, latest) and isn't simply a number (does semver
matchign against default remote).

Closes #37

Signed-off-by: Ryan Graham <r.m.graham@gmail.com>