optimization: memoize getRemoteVersionsAsync
This function is called multiple times by outdated() and when there are
lots of versions of node installed it can result in a lot of extra
http/https requests just to get the same data. Since it isn't likely to
change during the course of a single execution of nvs there's no reason
not to cache the results when we get them.