Add outdated command (#167)
* deps: add semver

Given the nature of nvs, it makes sense that this would eventually be
pulled in at some point. It may as well be now.

* cli: add 'outdated' command

Output is the same as 'nvs ls' except adds annotations to each version
if there is a newer version of that semver-major release.

* optimization: memoize getRemoteVersionsAsync

This function is called multiple times by outdated() and when there are
lots of versions of node installed it can result in a lot of extra
http/https requests just to get the same data. Since it isn't likely to
change during the course of a single execution of nvs there's no reason
not to cache the results when we get them.

* cli: enhance output, add test case

Display both patch and minor versions, in a more concise format.

Co-authored-by: Ryan Graham <r.m.graham@gmail.com>