Correctly wait on promisified cmdShim
The internal cmd-shim module of npm has been promisified in
https://github.com/npm/cmd-shim/commit/48fd8118057a19830e154231991a9acdb1a5610d
(Version 4+ and above I think)

As such the callback was never invoked, and the promise never resolved,
leading to code after it to not run, such as the code to write the
version label to an `.nvs` file. This causes the issue in #218.

Fixes #218